requirements.txt
aiohttp~=3.11.18
astral~=3.2
dotenv~=0.9.9
ecdsa~=0.19.1
llvmlite~=0.44.0
opencv-python~=4.11.0.86
pillow~=11.2.1
pwntools~=4.14.1
py7zr~=1.0.0
pycryptodome~=3.23.0
PyMuPDF~=1.25.5
python-dotenv~=1.1.0
pytz~=2025.2
rarfile~=4.2
requests~=2.32.3
rich~=14.0.0
sympy~=1.14.0
torch~=2.7.1
torchvision~=0.22.1
tqdm~=4.67.1
